moved location of yes and no boxes
I moved the location of the yes and no boxes, so they did not block the yes and no text. I also moved where the user needs to click in order to select yes or no, so the user is still required to click inside the box to select yes or no.